slice 2개를 restaurant slice 하나로 합쳐라
