+++ src/components/RegionListContainer.jsx
@@ -1,16 +1,16 @@
 import { useDispatch, useSelector } from 'react-redux';
 import { loadRestaurants } from '../redux_module/asyncActions';
 
-import { selectRegion } from '../redux_module/selectedSlice';
+import { selectRegion } from '../redux_module/RestaurantSlice';
 import ButtonList from './ButtonList';
 
 export default function RegionListContainer() {
   const dispatch = useDispatch();
 
-  const regions = useSelector((state) => state.groups.regions);
-  const category = useSelector((state) => state.selected.category);
+  const regions = useSelector((state) => state.restaurant.regions);
+  const category = useSelector((state) => state.restaurant.selected.category);
 
-  const selectedRegionName = useSelector((state) => state.selected.region);
+  const selectedRegionName = useSelector((state) => state.restaurant.selected.region);
 
   const handleClick = (region) => {
     dispatch(selectRegion(region));
